requirements.txt
names~=0.3.0
curl_cffi~=0.12.0
pytz~=2024.1
loguru~=0.7.2
httpx~=0.27.0
urllib3~=2.2.2
art~=6.2
PyYAML~=6.0.2
pydantic~=2.8.2
inquirer~=3.3.0
colorama~=0.4.6
better_proxy
tortoise-orm
imap_tools
aiofiles~=24.1.0
aiocsv~=1.3.2
rich~=13.9.2
python_socks
asyncpg~=0.30.0
psutil~=7.0.0
aiohttp~=3.11.18
eth-account~=0.13.7
eth-typing~=5.2.1
web3~=7.11.0
imap-tools~=1.10.0
python-socks~=2.7.1
base58~=2.1.1
cryptography~=45.0.5
openpyxl~=3.1.5
eth_abi~=5.2.0